requirements.txt
anthropic>=0.21.0
arxiv>=2.1.0
python-dotenv>=1.0.0
